REF: Move deployment package into cloud package
